+++ src/main/java/com/programmers/vouchermanagement/utils/Command.java
@@ -1,5 +1,7 @@
 package com.programmers.vouchermanagement.utils;
 
+import com.programmers.vouchermanagement.exception.CommandNotFoundException;
+
 import java.util.Arrays;
 
 public enum Command {
